+++ actionpack/test/dispatch/host_authorization_test.rb
@@ -221,15 +221,15 @@ class HostAuthorizationTest < ActionDispatch::IntegrationTest
     assert_match "Blocked host: www.example.com", response.body
   end
 
-  test "only compare to valid hostnames" do
+  test "blocks requests with invalid hostnames" do
     @app = ActionDispatch::HostAuthorization.new(App, ".example.com")
 
     get "/", env: {
-      "HOST" => "example.com#sub.example.com",
+      "HOST" => "attacker.com#x.example.com",
     }
 
     assert_response :forbidden
-    assert_match "Blocked host: example.com#sub.example.com", response.body
+    assert_match "Blocked host: attacker.com#x.example.com", response.body
   end
 
   test "config setting action_dispatch.hosts_response_app is deprecated" do