+++ actionview/test/template/compiled_templates_test.rb
@@ -52,7 +52,10 @@ class CompiledTemplatesTest < ActiveSupport::TestCase
   end
 
   def test_template_with_instance_variable_identifier
-    assert_equal "bar", render(template: "test/render_file_instance_variable", locals: { "@foo": "bar" })
+    expected_deprecation = "In Rails 7.1, @foo will be ignored."
+    assert_deprecated(expected_deprecation) do
+      assert_equal "bar", render(template: "test/render_file_instance_variable", locals: { "@foo": "bar" })
+    end
   end
 
   def test_template_gets_recompiled_when_using_different_keys_in_local_assigns